toggle between edit and read only
