+++ src/components/Profile/EditProfileInfo.scss
@@ -1,0 +1,13 @@
+.profile-info {
+
+  &_image {
+    >img{
+      height: 6em;
+      border-radius: 50%;
+    }
+
+    input[type='file'] {
+      color: rgba(0, 0, 0, 0)
+    }
+  }
+}
