+++ src/components/Profile/Profile.tsx
@@ -1,12 +1,31 @@
-import React from 'react';
+import React, {useState, useContext, useEffect} from 'react';
 import './Profile.scss';
-import ProfileInfo from './ProfileInfo'
+import EditProfileInfo from './EditProfileInfo'
+import ProfileInfo from './ProfileInfo';
+import { AuthContext } from '../../Auth';
+import { getDoc } from '../FirebaseApp';
+import app from '../FirebaseApp';
 
 function Profile() {
+  const { currentUser } = useContext(AuthContext);
+  const [editMode, setEditMode] = useState<boolean>(false);
+  const [profileInformation, setProfileInformation] = useState<any>()
+
+  useEffect(() => {
+    getDoc(app.auth().currentUser).then(data => {
+      setProfileInformation(data);
+    });
+  }, []);
+
   return (
     <>
     <div className='profile'>
-      <ProfileInfo/>
+      <button onClick={()=>setEditMode(!editMode)}>Edit</button>
+      {editMode ? 
+        <EditProfileInfo profileInfo={profileInformation}/>
+        :
+        <ProfileInfo profileInfo={profileInformation}/>
+      }
     </div>
     </>
   );
