--- src/components/Profile/ProfileInfo.scss
@@ -1,13 +1,0 @@
-.profile-info {
-
-  &_image {
-    >img{
-      height: 6em;
-      border-radius: 50%;
-    }
-
-    input[type='file'] {
-      color: rgba(0, 0, 0, 0)
-    }
-  }
-}
